DashBoard/requirements.txt
asgiref
Django
djangorestframework
mysqlclient